attempt to add fdk support
I think the docs are pretty stale on flow in this repo, was hoping to get
around to nailing this down today but didn't quite get there. anyway, posting
work to see if this is headed in the right direction, I mostly wanted to play
with flow stuff and check out the go flow library and I'd imagine moving to
use the fdk is something we want to do for this flow library.

I'm kind of unsure on what the cut point is for output. I get that using the
default format will just mop up everything to stdout and use it as output, but
I'm not sure what the cut point is for what we want the output to be vs. what
is additional output. As a start, this shoves everything into the fdk's
output, and takes input back through the fdk's input. this should allow us to
use json/http/default seamlessly but I'm not sure if stuff is garbled or not,
datum seemed to write some stuff directly to stdout as well as the function
provided to WithFlow, it seems like we possibly want to pass a buffer to
WithFlow inner function to buffer that and then add it to the flow output? but
I'm not entirely sure at the moment.

tried to make a minimal cut and just re-use the codec constructs to get
something working. anyway, I can continue tinkering here, possibly with some
guidance, or somebody else can take over.