use fdk instead of custom fn handler
this updates flow to use fdk for handling function framing and header
extraction. this allows flow to work with hot functions (and, when we update
to the new container format, presumably that). aside from being a good idea
from an abstraction level, the speed up is nice. the example went from 2.5s on
my laptop to 220ms so about 10x faster for a trivial flow (as expected).

this re-uses the psuedo-fdk codec interface just for ease of moving things
over. it seems wise to add tests for this to ease the migration process in the
future, i'm hoping to get to this separately rest of week and at that time can
try to do some tidying up. I did test that it works with http/json/default, it
took some tinkering but made it and it seems to work the same. bonus of using
fdk is that it's easy to add functions people made already to flows, woo.

if you want to test, you have to update (examples/hello-flow/Gopkg.toml) to
this branch and then run `dep ensure` before `fn deploy` so that it will
vendor the updated flow library to test it. I set that file to point to
flow@master so that when this gets merged it 'just works'.